Improve ControllableClock to leverage existing clocks in JDK
Instead of keeping a fixed time state itself, it instead simply uses a
delegate Clock which may be changed through a "test-friendly" API, i.e.
one can set specific time, simulate time passing, etc. Whether the time
is fixed or not depends on the underlying clock.

The constructor taking specific times (Instant, LocalDateTime, etc)
implicitly creates a fixed Clock, to not break existing use
of ControllableClock. To create a "unfixed" clock, simply pass a
system-clock to a constructor, and optionally set a new time if
needed. A regular self-ticking clock can be freezed with .freeze() or
using .set(..) and passing a fixed clock.