OneTimeAssignment treats null as any other assignment
Fixes a subtle bug where assigning a null was not considered an actual
assignment. The OneTimeAssignment is now capable of distinguishing if it
is not yet assigned, or if it has indeed been assigned null. This
required to break the type safety internally of the class, though this is
not exposed outside the class.