Add multituple collector
A multituple is similar to a multimap in that it consists of one first value and a List of values as the second tuple-value. This collector will collect tuples _where it is expected that all the first tuple-elements are all equal_ and rearrange them by putting the distinct first element into a new Tuple, and collate each of the second elements into a new List which is set as the second element of the new Tuple. If non-distinct values of the first elements of the tuples are encountered, an exception is thrown when executing the collector.