Make the primitive mappers yield null for SQL NULLs
The primitive mappers getInt, getByte, getDouble and so on in Mappers
did a naive wrapping of the call to its counterpart in ResultSet with no
check on ResultSet.wasNull(). As these mappers yield the boxed primitive
objects, the Mappers now correctly yield null in case of SQL NULL. 

Also added javadoc for these mappers, suggesting the use of the
getNullableXyz counterpart in case of mapping nullable columns.