Have a clean, extendible clock adjuster API
The previous release introduced a package scoped return type in the
TimeControllable interface, effectively making it not implementable
outside of the no.digipost.time package. While you would rarely want to
implement this, it is not good style to offer an interface which is not implementable.
Redesigning the interfaces and separating the default implementations
into separate interfaces, specifically for java.time.Clock subclasses,
and keeping the ClockAdjuster interface free from any implementation,
and, again, _only_ containing time adjusting methods.

The controllable clock API consists of 2 types: ControllableClock and
ClockAdjuster. The interfaces containing default implementations are
kept package private as for now, as they are only used by
ControllableClock to not clutter its implementation.