Support for threadsafe at-most-once execution
Two method overloads of OneTimeToggle.nowAndUnlessAlreadyToggled which
will ensure that a Runnable or Supplier is run at most one time, even if
several threads are trying to. Test is really stressing the execution of
this code to ensure this. I have also tried to implement the naïve
non-threadsafe version of this to verify that the test actually detects
this.