Add convertor method to all tuples
A tuple, e.g. a 2-ary Tuple, can be converterted to a an arbitrary
instance using tuple.to((v1, v2) -> new Something(v1, v2), or more
succinct: tuple.to(Something::new). This is a convenience API which
eliminates the need to "manually extract" each value of the tuple, and
instead either freely name the values with a lambda, or directly have
the values passed to a method reference with the equivalent parameter
signature as the parameterized types of the tuple.