Allow freezing time to truncated current instant (#19)
It may be desireable to freeze a clock to a lesser precision than
offered by the JDK, as the instant may be transformed and passed through facilities (e.g. databases)
which needs to decrease the precision. This will make it more convenient
to use freezed clocks in tests which also employ precise evaluating of
instances, but loss of precision is not relevant for regular freely
running system clock. E.g. queries to pick out rows with timestamp
before or after a given timestamp.