Add InputStreamIterator
This iterator wrapps an InputStream and reads chunks
from it returning the chunks in next() if i has more chunks.

I added two constructors. One that takes the chunk size for bytes
and one that uses DataSize. With DataSize bytes or kiloBytes
only should be used, since too big chunks would defeat the purpose
of this iterator. Its better to read it fully to a byte[].
DataSize uses long, and thus numbers bigger that Integer.MAX will
result in a failing iterator.